Add test for runLauncher (#2629)

+++ cmd/launcher/launcher.go
@@ -12,6 +12,7 @@ import (
 	"runtime"
 	"strconv"
 	"strings"
+	"sync"
 	"time"
 
 	"github.com/apache/thrift/lib/go/thrift"
@@ -84,6 +85,13 @@ const (
 	localizationsSubsystemName            = "localizations"
 )
 
+// setThriftInterval ensures we modify the thrift.ServerConnectivityCheckInterval global variable
+// only once. It exists largely to avoid data races in tests -- in production usage, we only ever
+// expect it to be called once.
+var setThriftInterval = sync.OnceFunc(func() {
+	thrift.ServerConnectivityCheckInterval = 100 * time.Millisecond
+})
+
 // runLauncher is the entry point into running launcher. It creates a
 // rungroups with the various options, and goes! If autoupdate is
 // enabled, the finalizers will trigger various restarts.
@@ -91,7 +99,7 @@ func runLauncher(ctx context.Context, cancel func(), multiSlogger, systemMultiSl
 	initialTraceBuffer := exporter.NewInitialTraceBuffer()
 	ctx, startupSpan := observability.StartSpan(ctx)
 
-	thrift.ServerConnectivityCheckInterval = 100 * time.Millisecond
+	setThriftInterval()
 
 	logger := ctxlog.FromContext(ctx)
 	logger = log.With(logger, "caller", log.DefaultCaller, "session_pid", os.Getpid())
@@ -173,9 +181,11 @@ func runLauncher(ctx context.Context, cancel func(), multiSlogger, systemMultiSl
 		return fmt.Errorf("detecting platform: %w", err)
 	}
 
-	debugAddrPath := filepath.Join(rootDirectory, "debug_addr")
-	debug.AttachDebugHandler(debugAddrPath, slogger)
-	defer os.Remove(debugAddrPath)
+	if opts.Debug {
+		debugAddrPath := filepath.Join(rootDirectory, "debug_addr")
+		debug.AttachDebugHandler(debugAddrPath, slogger)
+		defer os.Remove(debugAddrPath)
+	}
 
 	// open the database for storing launcher data, we do it here
 	// because it's passed to multiple actors. Add a timeout to
@@ -237,8 +247,12 @@ func runLauncher(ctx context.Context, cancel func(), multiSlogger, systemMultiSl
 	gowrapper.Go(ctx, slogger, func() {
 		// Wait a little bit before adding exclusions -- some osquery files get created right after
 		// startup and we want to let that settle before handling exclusions.
-		time.Sleep(2 * time.Minute)
-		timemachine.AddExclusions(ctx, k)
+		select {
+		case <-ctx.Done():
+			// launcher shut down before our sleep finished -- nothing to do here but exit.
+		case <-time.After(2 * time.Minute):
+			timemachine.AddExclusions(ctx, k)
+		}
 	})
 
 	if k.Debug() && runtime.GOOS != "windows" {
